# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Desktop AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List chat sessions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{id}:
    get:
      operationId: getChatSession
      tags: [chat]
      summary: Get a chat session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionWithMessages"
        "404":
          description: Not found
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chat/sessions/{id}/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response with optional action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /activity:
    get:
      operationId: listActivityLog
      tags: [activity]
      summary: List activity log entries
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [chat, file_op, scheduled, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
  /activity/{id}/undo:
    post:
      operationId: undoActivity
      tags: [activity]
      summary: Undo a file operation from the activity log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Undo result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UndoResult"
...
  /ai/models/pull:
    post:
      operationId: pullAiModel
      tags: [ai]
      summary: Pull a model via Ollama
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelPullInput"
      responses:
        "200":
          description: Pull initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPullStatus"
  /ai/models/{name}:
    delete:
      operationId: deleteAiModel
      tags: [ai]
      summary: Delete a local Ollama model
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── File Operations ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ChatSession:
      type: object
      required: [id, title, createdAt, updatedAt, messageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
    ChatSessionInput:
      type: object
      properties:
        title:
          type: string
    ChatSessionWithMessages:
      type: object
      required: [id, title, createdAt, updatedAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        actionType:
          type: ["string", "null"]
        actionData:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          $ref: "#/components/schemas/ChatMessage"
        action:
          $ref: "#/components/schemas/AiAction"
    AiAction:
      type: object
      required: [type, description]
      properties:
        type:
          type: string
          enum: [file_op, shell, info, none]
        description:
          type: string
        payload:
          type: ["string", "null"]
        requiresConfirmation:
          type: boolean
...
    ActivityEntry:
      type: object
      required: [id, type, description, createdAt, undoable]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [chat, file_op, scheduled]
        description:
          type: string
        detail:
          type: ["string", "null"]
        filesAffected:
          type: array
          items:
            type: string
        undoable:
          type: boolean
        undone:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UndoResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        filesRestored:
          type: array
          items:
            type: string
...
    ModelPullInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ModelPullStatus:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
        progress:
          type: ["number", "null"]

    # ── Files ────────────────────────────────────────────────────────────────
...