# lib/api-zod/src/generated/types/modelPullStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Desktop AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelPullStatus {
  name: string;
  status: string;
  /** @nullable */
  progress?: number | null;
}